Fixed heartbeats project paths.Added base reservations project
