+++ ChargingStation.Backend/API/ChargingStation.Heartbeats/Models/Request/GetHeartbeatRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace ChargingStation.Heartbeats.Models.Request;
+
+public class GetHeartbeatRequest
+{
+    public required string PartitionKey { get; set; }
+    public required string RowKey { get; set; }
+}
