+++ ChargingStation.Backend/API/ChargingStation.Heartbeats/Properties/launchSettings.json
