+++ ChargingStation.Backend/API/ChargingStation.Heartbeats/Services/ChargePoint/IChargePointHttpService.cs
@@ -1,0 +1,6 @@
+﻿namespace ChargingStation.Heartbeats.Services.Connectors;
+
+public interface IChargePointHttpService
+{
+    Task<bool> GetByIdAsync(string chargePointId, CancellationToken cancellationToken = default);
+}
