+++ ChargingStation.Backend/API/ChargingStation.Heartbeats/Services/HeartbeatService/HeartbeatService.cs
@@ -6,23 +6,23 @@ using ChargingStation.Heartbeats.Models.Request;
 using ChargingStation.Heartbeats.Services.Connectors;
 using ChargingStation.Infrastructure.AzureTableStorage;
 
-namespace ChargingStation.Heartbeats.Services.HeartbeatService
-{
-    public class HeartbeatService : IHeartbeatService
-    {
-        private readonly IChargePointHttpService _chargePointService;
-        private readonly ITableManager<HeartbeatEntity> _tableManager;
-        private readonly string _tableName;
+namespace ChargingStation.Heartbeats.Services.HeartbeatService;
 
-        public HeartbeatService(IChargePointHttpService chargePointService, ITableManager<HeartbeatEntity> tableManager, IConfiguration configuration)
-        {
+public class HeartbeatService : IHeartbeatService
+{
+    private readonly IChargePointHttpService _chargePointService;
+    private readonly ITableManager<HeartbeatEntity> _tableManager;
+    private readonly string _tableName;
+
+    public HeartbeatService(IChargePointHttpService chargePointService, ITableManager<HeartbeatEntity> tableManager, IConfiguration configuration)
+    {
             _chargePointService = chargePointService;
             _tableManager = tableManager;
             _tableName = configuration.GetConnectionString("AzureTableStorage:HeartbeatTable")!;
         }
 
-        public async Task AddHeartbeatAsync(HeartbeatEntity request, CancellationToken cancellationToken = default)
-        {
+    public async Task AddHeartbeatAsync(HeartbeatEntity request, CancellationToken cancellationToken = default)
+    {
             var chargePoint = await _chargePointService.GetByIdAsync(request.PartitionKey, cancellationToken);
 
             if (!chargePoint)
@@ -34,8 +34,8 @@ namespace ChargingStation.Heartbeats.Services.HeartbeatService
             await _tableManager.AddEntityAsync(_tableName, request);
         }
 
-        public async Task<HeartbeatEntity> GetByIdAsync(GetHeartbeatRequest request, CancellationToken cancellationToken = default)
-        {
+    public async Task<HeartbeatEntity> GetByIdAsync(GetHeartbeatRequest request, CancellationToken cancellationToken = default)
+    {
             var heartbeat = await _tableManager.GetEntityAsync(_tableName, request.PartitionKey, request.RowKey);
 
             if (heartbeat is null)
@@ -44,8 +44,8 @@ namespace ChargingStation.Heartbeats.Services.HeartbeatService
             return heartbeat;
         }
 
-        public async Task<List<HeartbeatEntity>> GetAsync(CancellationToken cancellationToken = default)
-        {
+    public async Task<List<HeartbeatEntity>> GetAsync(CancellationToken cancellationToken = default)
+    {
             var heartbeats = await _tableManager.GetAllEntitiesAsync(_tableName);
 
             if (!heartbeats.Any())
@@ -54,8 +54,8 @@ namespace ChargingStation.Heartbeats.Services.HeartbeatService
             return heartbeats;
         }
 
-        public async Task<HeartbeatResponse> ProcessHeartbeatAsync(HeartbeatRequest request, Guid chargePointId, CancellationToken cancellationToken = default)
-        {
+    public async Task<HeartbeatResponse> ProcessHeartbeatAsync(HeartbeatRequest request, Guid chargePointId, CancellationToken cancellationToken = default)
+    {
             DateTimeOffset currentTime = DateTimeOffset.UtcNow;
 
             var response = new HeartbeatResponse(currentTime);
@@ -65,5 +65,4 @@ namespace ChargingStation.Heartbeats.Services.HeartbeatService
 
             return response;
         }
-    }
 }
