--- ChargingStation.Heartbeats/Models/Request/GetHeartbeatRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace ChargingStation.Heartbeats.Models.Request
-{
-    public class GetHeartbeatRequest
-    {
-        public required string PartitionKey { get; set; }
-        public required string RowKey { get; set; }
-    }
-}
