--- ChargingStation.Heartbeats/Services/ChargePoint/IChargePointHttpService.cs
@@ -1,7 +1,0 @@
-﻿namespace ChargingStation.Heartbeats.Services.Connectors
-{
-    public interface IChargePointHttpService
-    {
-        Task<bool> GetByIdAsync(string chargePointId, CancellationToken cancellationToken = default);
-    }
-}
